Reverts partial support for nested OpenMP in Intel KMPC runtime.
The helper structures used to store thread data were stored in a matrix indexed by thread id and nesting level.
But this is not enough, beyond the 2nd nested level, multiple children threads have the same parent id.
In order to distinguish the parent, the whole chain of ancestor id's has to be queried, and the structure
holding the helper data should be a tree.

As a collateral, this structure had many problems initializing as it would allocate slots for omp_get_max_threads(),
but some runtimes crash querying this before it is initialized (ARM), or return a wrong value of 1 (Nanos).

A workaround to keep using this helper structure would include to add mutexes in all accesses to it to allow
reallocating dynamically if the thread using it doesn't have slots allocated, but would still fail for higher
degrees of nested parallelism and add significant overhead.

So we decide to revert to the previous situation. The outlined routine is stored in a global variable from the master
thread, and queried from the children threads. This only supports 1 level of parallelism, but at least doesn't
crash the initializations.

Change-Id: I31b4d27815f749983ea6f5d416392b7560cef973