Fix missing library libsymLite.so
Since dyninst/dyninst@739ad57883f032fb18a6c741f0c2f98e9a766bc7, the
library can be built with `LIGHTWEIGHT_SYMTAB=ON` which allows (or
disallows) for the building of the lightweight version of the dyninst
library (particularly the Stackwalking, Parsing and ProcControl.

This commit takes that in account by checking whether the light version
is available, and if not, adds the relevant shared libraries.