Refactors calls to Enter_Instrumentation and Leave_Instrumentation in all wrappers
Moves the Enter_Instrumentation and Leave_Instrumentation to be the first and last instructions to be executed when entering/leaving the wrappers, as well as the API call of Extrae_init and the library constructor. This forbids the tracing of unwanted inner calls when already in instrumentation or during initialization.